Move post processor to decoders
