+++ qiskit_ibm_runtime/decoders/executor_estimator/__init__.py
@@ -10,7 +10,4 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
-"""V0.1 post processor for executor-based EstimatorV2"""
-
-# Import in order to register.
-from .post_processor_v0_1 import estimator_v2_post_processor_v0_1
+"""Result decoder for executor-based Estimator."""
