+++ qiskit_ibm_runtime/executor_estimator/__init__.py
@@ -11,5 +11,3 @@
 # that they have been altered from the originals.
 
 """Executor-based EstimatorV2"""
-
-from .post_processors.registry import ESTIMATOR_POST_PROCESSORS
